+++ package.json
@@ -15,6 +15,7 @@
   "dependencies": {
     "chai": "^4.2.0",
     "cheerio": "^1.0.0-rc.2",
+    "ejs": "^2.6.1",
     "express": "^4.16.3",
     "mocha": "^5.2.0",
     "supertest": "^3.3.0"
