+++ views/index.ejs
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title></title>
+</head>
+<body>
+    <h1>Mars Index</h1>
+    <ul>
+        <% for(let i = 0; i < marsMissions.length; i++) {%>
+            <li>
+                <a href='/missions/<%= i %>'><%= marsMissions[i].name %></a>
+            </li>
+        <% }%>
+    </ul>
+</body>
+</html>
