+++ views/show.ejs
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html>
+<head>
+	<title></title>
+</head>
+<body>
+
+<h1>Missions Show Page</h1>
+
+The <%= mission.name %> Lanches on <%= mission.launchDate %>,  Operated by <%= mission.operator %> . The type of mission will be a <%=mission.missionType%> <%=mission.img%>
+
+<a href="http://localhost:3000/missions">go back</a>
+</body>
+</html>
